requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
openai>=1.0.0
openai==1.6.1
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
werkzeug==3.0.1
